Build instructions `x86_64-unknown-linux-fortanixvme` target

+++ src/ci/docker/scripts/musl-toolchain.sh
@@ -34,7 +34,7 @@ TARGET=$ARCH-linux-musl
 LINUX_HEADERS_SITE=https://ci-mirrors.rust-lang.org/rustc/sabotage-linux-tarballs
 LINUX_VER=headers-4.19.88
 
-OUTPUT=/usr/local
+OUTPUT="${INSTALL_DIR:=/usr/local}"
 shift
 
 # Ancient binutils versions don't understand debug symbols produced by more recent tools.
@@ -43,7 +43,10 @@ shift
 # debuggers can't walk the stack, etc. Fixes #90103.
 export CFLAGS="-fPIC -g1 $CFLAGS"
 
-git clone https://github.com/richfelker/musl-cross-make # -b v0.9.9
+if [ ! -d "musl-cross-make" ]; then
+    git clone https://github.com/richfelker/musl-cross-make # -b v0.9.9
+fi
+
 cd musl-cross-make
 # A version that includes support for building musl 1.2.3
 git checkout fe915821b652a7fa37b34a596f47d8e20bc72338
@@ -53,16 +56,24 @@ hide_output make install TARGET=$TARGET MUSL_VER=1.2.3 LINUX_HEADERS_SITE=$LINUX
 
 cd -
 
-# Install musl library to make binaries executable
-ln -s $OUTPUT/$TARGET/lib/libc.so /lib/ld-musl-$ARCH.so.1
-echo $OUTPUT/$TARGET/lib >> /etc/ld-musl-$ARCH.path
+if [ "${USER}" == "root" ]; then
+    # Install musl library to make binaries executable
+    if [ ! -f "/lib/ld-musl-$ARCH.so.1" ]; then
+        ln -s $OUTPUT/$TARGET/lib/libc.so /lib/ld-musl-$ARCH.so.1
+    fi
+    echo $OUTPUT/$TARGET/lib >> /etc/ld-musl-$ARCH.path
 
-# Now when musl bootstraps itself create proper toolchain symlinks to make build and tests easier
-if [ "$REPLACE_CC" = "1" ]; then
-    for exec in cc gcc; do
-        ln -s $TARGET-gcc /usr/local/bin/$exec
-    done
-    for exec in cpp c++ g++; do
-        ln -s $TARGET-g++ /usr/local/bin/$exec
-    done
+    # Now when musl bootstraps itself create proper toolchain symlinks to make build and tests easier
+    if [ "$REPLACE_CC" = "1" ]; then
+       for exec in cc gcc; do
+            if [ ! -f "/usr/local/bin/$exec" ]; then
+                ln -s $TARGET-gcc /usr/local/bin/$exec
+            fi
+        done
+        for exec in cpp c++ g++; do
+            if [ ! -f "/usr/local/bin/$exec" ]; then
+                ln -s $TARGET-g++ /usr/local/bin/$exec
+            fi
+        done
+    fi
 fi
